packages/vgplot/widget/pyproject.toml
[project]
name = "mosaic-widget"
description = "A Jupyter widget for Mosaic"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anywidget>=0.9.0",
  "duckdb>=1.3.1",
  "narwhals>=1.46.0",
  "pyarrow",
]

[project.urls]
homepage = "https://github.com/uwdata/mosaic"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
only-packages = true
artifacts = [
  "mosaic_widget/static/"
]

[dependency-groups]
dev = [
  "jupyterlab",
  "mypy",
  "pandas",
  "pyyaml",
  "ruff",
  "watchfiles",
]
test = [
  "dask",
  "ibis-framework",
  "modin",
  "polars",
  "pyarrow-hotfix",
  "pytest",
]

[tool.hatch.version]
path = "package.json"
pattern = "\"version\": \"(?P<version>.+?)\""

# https://mypy.readthedocs.io/en/stable/config_file.html
[tool.mypy]
files = "mosaic_widget/**/*.py"
strict = false
ignore_missing_imports = true
